Use random head selection when BKT selection fails
For reasons unclear to me, SPANN sometimes -- especially but not only with
distance=IP indexes -- fails to select heads.  More vectors helps, as does
more randomness among vectors.  One egregious case is 100 vectors that
still can't build an IP index in the MySQL test suite.

So as a workaround: if the data are too small and/or artificial to select
heads for an index, fall back on random head selection instead.